requirements.txt
chardet==5.2.0
pillow==10.3.0
PyPDF2==3.0.1
python-dotenv==1.0.1
reportlab==4.1.0
